Refactor - 'entering degrees' spec
This spec has grown quite long so let's split it into smaller specs
