+++ spec/system/candidate_interface/entering_details/degrees/candidate_entering_degree_with_missing_information_spec.rb
